+++ baybe/targets.py
@@ -1,10 +1,10 @@
 """
-Functionality for different type of targets
+Functionality for different type of targets.
 """
 
 from __future__ import annotations
 
-from abc import ABC
+from abc import ABC, abstractmethod
 from typing import ClassVar, Dict, List, Literal, Optional, Tuple
 
 import numpy as np
@@ -54,10 +54,14 @@ class Target(ABC, BaseModel, extra=Extra.forbid):
         check_if_in(param_type, list(Target.SUBCLASSES.keys()))
         return cls.SUBCLASSES[param_type](**config)
 
+    @abstractmethod
+    def transform(self, data: pd.DataFrame) -> pd.DataFrame:
+        """Creates a new target representation matching the given specifications."""
+
 
 class NumericalTarget(Target):
     """
-    Class for numerical targets
+    Class for numerical targets.
     """
 
     type = "NUM"