+++ baybe/utils.py
@@ -1,15 +1,30 @@
 """
 Collection of small utilities
 """
-
 from __future__ import annotations
 
 from functools import partial
-from typing import Any, Dict, Iterable, Optional, Tuple, TYPE_CHECKING, Union
+
+from typing import (
+    Any,
+    Dict,
+    Iterable,
+    List,
+    Optional,
+    Tuple,
+    Type,
+    TYPE_CHECKING,
+    Union,
+)
+
+from urllib.request import urlopen
 
 import numpy as np
 import pandas as pd
 import torch
+from mordred import Calculator, descriptors
+from rdkit import Chem, RDLogger
+from rdkit.Chem.AllChem import GetMorganFingerprintAsBitVect
 from torch import Tensor
 
 if TYPE_CHECKING:
@@ -18,7 +33,7 @@ if TYPE_CHECKING:
 
 def is_valid_smiles(smiles: str) -> bool:
     """
-    Test if a SMILEs string is valid according to RDKit.
+    Test if a SMILES string is valid according to RDKit.
 
     Parameters
     ----------
@@ -28,7 +43,11 @@ def is_valid_smiles(smiles: str) -> bool:
     -------
         bool, True if smiles is valid, False else
     """
-    raise NotImplementedError("This function is not implemented yet")
+    try:
+        mol = Chem.MolFromSmiles(smiles)
+        return mol is not None
+    except Exception:
+        return False
 
 
 def to_tensor(*dfs: pd.DataFrame) -> Union[Tensor, Iterable[Tensor]]:
@@ -48,6 +67,36 @@ def check_if_in(element: Any, allowed: list):
         raise ValueError(
             f"The value '{element}' is not allowed. Must be one of {allowed}."
         )
+
+
+def name_to_smiles(name: str) -> str:
+    """
+    Convert from chemical name to SMILES string using chemical identifier resolver.
+
+    Parameters
+    ----------
+    name : str
+        Name or nickname of compound.
+
+    Returns
+    ----------
+    str
+        SMILES string corresponding to chemical name.
+    """
+
+    name = name.replace(" ", "%20")
+
+    try:
+        url = "http://cactus.nci.nih.gov/chemical/structure/" + name + "/smiles"
+        with urlopen(url) as web:
+            smiles = web.read().decode("utf8")
+        smiles = str(smiles)
+        if "</div>" in smiles:
+            return ""
+
+        return smiles
+    except Exception:
+        return ""
 
 
 def add_fake_results(
@@ -197,3 +246,261 @@ def add_noise(
                     f"Parameter noise_type was {noise_type} but must be either "
                     f'"absolute" or "relative_percent"'
                 )
+
+
+def smiles_to_mordred_features(
+    smiles_list: list, prefix: str = "", dropna: bool = True
+) -> pd.DataFrame:
+    """
+    Compute Mordred chemical descriptors for a list of SMILES strings.
+
+    Parameters
+    ----------
+    smiles_list : list
+        List of SMILES strings.
+    prefix : str
+        Name prefix for each descriptor (e.g., nBase --> name_nBase).
+    dropna : bool
+        If true, drop columns which contain np.NaNs.
+
+    Returns
+    ----------
+    pandas.DataFrame
+        DataFrame containing overlapping Mordred descriptors for each SMILES
+        string.
+    """
+    calc = Calculator(descriptors)
+
+    output = []
+    for smiles in smiles_list:
+        try:
+            data_i = calc(Chem.MolFromSmiles(smiles)).fill_missing()
+        except Exception:
+            data_i = np.full(len(calc.descriptors), np.NaN)
+
+        output.append(list(data_i))
+
+    descriptor_names = list(calc.descriptors)
+    columns = []
+    for entry in descriptor_names:
+        columns.append(prefix + str(entry))
+
+    dataframe = pd.DataFrame(data=output, columns=columns)
+
+    if dropna:
+        dataframe = dataframe.dropna(axis=1)
+
+    return dataframe
+
+
+def smiles_to_rdkit_features(
+    smiles_list: list, prefix: str = "", dropna: bool = True
+) -> pd.DataFrame:
+    """
+    Compute RDKit chemical descriptors for a list of SMILES strings.
+
+    Parameters
+    ----------
+    smiles_list : list
+        List of SMILES strings.
+    prefix : str
+        Name prefix for each descriptor (e.g., nBase --> name_nBase).
+    dropna : bool
+        If true, drop columns which contain np.NaNs.
+
+    Returns
+    ----------
+    pandas.DataFrame
+        DataFrame containing overlapping RDKit descriptors for each SMILES
+        string.
+    """
+
+    mols = []
+    for smiles in smiles_list:
+        try:
+            mol = Chem.MolFromSmiles(smiles)
+            if mol is None:
+                raise ValueError(
+                    f"The smiles {smiles} does not seem " f"to chemically valid"
+                )
+            mols.append(mol)
+        except Exception as ex:
+            raise ValueError(
+                f"The smiles {smiles} does not seem " f"to chemically valid"
+            ) from ex
+
+    res = []
+    for mol in mols:
+        desc = {prefix + dname: func(mol) for dname, func in Chem.Descriptors.descList}
+        res.append(desc)
+
+    df = pd.DataFrame(res)
+    if dropna:
+        df = df.dropna(axis=1)
+
+    return df
+
+
+def smiles_to_fp_features(
+    smiles_list: list,
+    prefix: str = "",
+    dtype: Union[Type[int], Type[float]] = int,
+    radius: int = 4,
+    n_bits: int = 1024,
+) -> pd.DataFrame:
+    """
+    Compute standard morgan molecule fingerprints for a list of SMILES strings.
+
+    Parameters
+    ----------
+    smiles_list : list
+        List of SMILES strings.
+    prefix : str
+        Name prefix for each descriptor (e.g., nBase --> name_nBase).
+    dtype : datatype int or float
+        Specifies whether fingerprints will have int or float datatype
+    radius : int
+        Radius for the Morgan Fingerprint
+    n_bits : int
+        Number of bits for the Morgan fingerprint
+
+    Returns
+    ----------
+    pandas.DataFrame
+        DataFrame containing Morgan fingerprints for each SMILES string.
+    """
+
+    mols = []
+    for smiles in smiles_list:
+        try:
+            mol = Chem.MolFromSmiles(smiles)
+            if mol is None:
+                raise ValueError(
+                    f"The smiles {smiles} does not seem " f"to chemically valid"
+                )
+            mols.append(mol)
+        except Exception as ex:
+            raise ValueError(
+                f"The smiles {smiles} does not seem " f"to chemically valid"
+            ) from ex
+
+    res = []
+    for mol in mols:
+        RDLogger.logger().setLevel(RDLogger.CRITICAL)
+
+        fingerp = GetMorganFingerprintAsBitVect(mol, radius, nBits=n_bits).ToBitString()
+        fingerp = map(int, fingerp)
+        fpvec = np.array(list(fingerp))
+        res.append(
+            {prefix + "FP" + f"{k + 1}": dtype(bit) for k, bit in enumerate(fpvec)}
+        )
+
+    df = pd.DataFrame(res)
+
+    return df
+
+
+def df_drop_single_value_columns(df: pd.DataFrame) -> pd.DataFrame:
+    """
+    Drop dataframe columns with zero variance. Adapted from edbo.
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        The dataframe to be cleaned
+    Returns
+    -------
+    pandas.DataFrame
+        A new dataframe
+    """
+    to_keep = []
+    for k in range(len(df.columns)):
+        if len(df.iloc[:, k].drop_duplicates()) > 1:
+            to_keep.append(df.columns.values[k])
+
+    return df[to_keep]
+
+
+def df_drop_string_columns(
+    df: pd.DataFrame, ignore_list: Optional[list] = None
+) -> pd.DataFrame:
+    """
+    Drop dataframe columns with non-numeric values. Adapted from edbo.
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        The dataframe to be cleaned
+    ignore_list : list
+        List of columns that should not be dropped, even if they include strings
+    Returns
+    -------
+    pandas.DataFrame
+        A new dataframe
+    """
+    ignore_list = ignore_list or []
+
+    to_keep = []
+    for k in range(len(df.columns)):
+        unique = df.iloc[:, k].drop_duplicates()
+        keep_q = True
+        for j in range(len(unique)):
+            if isinstance(unique.iloc[j], str):
+                keep_q = False
+                break
+        if keep_q:
+            to_keep.append(df.columns.values[k])
+
+    return df[to_keep + ignore_list]
+
+
+def df_uncorrelated_features(
+    df: pd.DataFrame, exclude_list: Optional[List[str]] = None, threshold: float = 0.7
+):
+    """
+    Returns an uncorrelated set of features. Adapted from edbo.
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        The dataframe to be cleaned
+    exclude_list : list of strings or None
+        If provided this defines the columns that should be ignored
+    threshold : float
+        Threshold for column-column correlation above which columns should be dropped
+    Returns
+    -------
+    data : pandas.DataFrame
+        A new dataframe
+    """
+
+    if exclude_list is None:
+        data = df.copy()
+    else:
+        data = df.drop(columns=exclude_list)
+
+    corr = data.corr().abs()
+    to_keep = []
+    for i in range(len(corr.iloc[:, 0])):
+        above = corr.iloc[:i, i]
+        if len(to_keep) > 0:
+            above = above[to_keep]
+        if len(above[above < threshold]) == len(above):
+            to_keep.append(corr.columns.values[i])
+
+    data = data[to_keep]
+
+    if exclude_list is not None:
+        data[exclude_list] = df.loc[:, exclude_list]
+
+    return data
+
+
+class StrictValidationError(Exception):
+    """
+    This class is used as a workaround that can be used to stop pydantic from continuing
+    validating other members when the validation of a specific field fails.
+
+    The issue is described here:
+    https://github.com/pydantic/pydantic/issues/3915
+    """